requirements.txt
dlib==19.22.99
Flask==1.1.2
Flask_WTF==1.0.1
hyperopt==0.2.7
imageio==2.9.0
moviepy==1.0.3
numpy==1.20.3
opencv_python==4.5.5.62
pandas==1.3.4
Pillow==9.1.1
psycopg2==2.9.3
scikit_image==0.18.3
skimage==0.0
tensorflow==2.8.0
tflearn==0.5.0
Werkzeug==2.0.2
WTForms==2.3.3
